refactor(bilibili): 调整B站解析器中音频质量枚举定义位置
将HookAudioQuality枚举类从bilibili子模块移至主模块，
统一在__init__.py中定义并覆盖bilibili_api.video.AudioQuality，
便于全局使用和维护音频清晰度配置
